clean: and suppress warnings

+++ HFM.Proteins.Tests/ArtifactFolder.cs
@@ -1,34 +1,36 @@
-﻿using System;
-using System.IO;
+﻿namespace HFM.Proteins;
 
-namespace HFM.Proteins
+public sealed class ArtifactFolder : IDisposable
 {
-    public class ArtifactFolder : IDisposable
+    public string Path { get; }
+
+    public ArtifactFolder() : this(Environment.CurrentDirectory)
     {
-        public string Path { get; private set; }
 
-        public ArtifactFolder()
-           : this(Environment.CurrentDirectory)
+    }
+
+    public ArtifactFolder(string basePath)
+    {
+        Path = GetRandomPath(basePath);
+        Directory.CreateDirectory(Path);
+    }
+
+    public string GetRandomFilePath() => GetRandomPath(Path);
+
+    public static string GetRandomPath(string path) => System.IO.Path.Combine(path, System.IO.Path.GetRandomFileName());
+
+    public void Dispose()
+    {
+        try
         {
-
-        }
-
-        public ArtifactFolder(string basePath)
-        {
-            Path = System.IO.Path.Combine(basePath, System.IO.Path.GetRandomFileName());
-            Directory.CreateDirectory(Path);
-        }
-
-        public void Dispose()
-        {
-            try
+            if (Directory.Exists(Path))
             {
                 Directory.Delete(Path, true);
             }
-            catch (Exception)
-            {
-
-            }
+        }
+        catch (Exception)
+        {
+            // do nothing
         }
     }
 }
